Update README with new docker deployment instructions
The README.md file has been revised to include more detailed instructions on how to utilize the new docker images for deployment. The focus lies on guiding users through the process of using the 'docker push' command with a specific tagname to push an updated version of the 'php-deployer' docker image for deployment.